First stab at native dialogues
