+++ src/Strem.Platforms.Windows/Strem.Platforms.Windows.csproj
@@ -9,7 +9,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="NativeFileDialogSharp" Version="0.5.0" />
       <PackageReference Include="NAudio" Version="2.2.1" />
       <PackageReference Include="System.Speech" Version="8.0.0" />
       <PackageReference Include="TextCopy" Version="6.2.1" />
